+++ src/main/java/chess/dao/MysqlConnector.java
@@ -1,0 +1,32 @@
+package chess.dao;
+
+import java.sql.Connection;
+import java.sql.DriverManager;
+import java.sql.SQLException;
+
+public class MysqlConnector {
+
+    private static final String URL = "jdbc:mysql://localhost:3307/chess";
+    private static final String USER = "user";
+    private static final String PASSWORD = "password";
+
+    public static Connection connect() {
+        Connection connection = null;
+        try {
+            connection = DriverManager.getConnection(URL, USER, PASSWORD);
+        } catch (final SQLException exception) {
+            exception.printStackTrace();
+        }
+        return connection;
+    }
+
+    static void close(Connection connection) {
+        try {
+            if (connection != null) {
+                connection.close();
+            }
+        } catch (SQLException e) {
+            e.printStackTrace();
+        }
+    }
+}
